[front/components/actions/mcp] - refactor: rename CreateMCPServerSheet to CreateMCPServerDialog
 - Renamed files and references from `CreateMCPServerSheet` to the new name `CreateMCPServerDialog` for clarity and consistency.
