+++ front/components/markdown/tool/ToolSetupCard.tsx
@@ -8,7 +8,7 @@
 import { Button, ContentMessage } from "@dust-tt/sparkle";
 import React, { useMemo, useState } from "react";
 
-import { CreateMCPServerDialog } from "@app/components/actions/mcp/CreateMCPServerSheet";
+import { CreateMCPServerDialog } from "@app/components/actions/mcp/CreateMCPServerDialog";
 import { getIcon } from "@app/components/resources/resources_icons";
 import type { InternalMCPServerNameType } from "@app/lib/actions/mcp_internal_actions/constants";
 import { isInternalMCPServerOfName } from "@app/lib/actions/mcp_internal_actions/constants";
